ml-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
tflite-runtime==2.14.0 ; sys_platform == 'linux'
tensorflow==2.15.0 ; sys_platform == 'win32'
shap==0.44.1
Pillow==10.2.0
numpy==1.26.4
boto3==1.34.50
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
scikit-image==0.22.0
matplotlib==3.8.3
